Code Refactoring + Add Documentation + Remove Unused Imports + Separate Widgets in Directions Page
