+++ lib/models/building.dart
@@ -1,14 +1,11 @@
-import 'package:flutter/material.dart';
+import 'package:concordia_navigation/storage/app_constants.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
+//This class is a building model, with building name, list of edges, and polygon.
 class Building {
-
   String name;
-
   List<LatLng> edges;
-
   bool updated = false;
-
   Polygon _outline;
 
   Building(String name, List<LatLng> edges) {
@@ -20,12 +17,11 @@ class Building {
     if (_outline == null || updated == false) {
       _outline = new Polygon(
         polygonId: PolygonId(name),
-        fillColor: Color(0xFF73C700).withOpacity(0.3),
+        fillColor: greenColor.withOpacity(0.3),
         consumeTapEvents: false,
         geodesic: false,
         points: edges,
         strokeWidth: 0,
-        strokeColor: Colors.transparent,
       );
     }
     return _outline;
