+++ lib/providers/buildings_data.dart
@@ -1,5 +1,5 @@
 import 'package:concordia_navigation/models/building.dart';
-import 'package:concordia_navigation/storage/campus_polygons.dart';
+import 'package:concordia_navigation/storage/campus_buildings.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:flutter/widgets.dart';
 
@@ -24,7 +24,7 @@ class BuildingsData extends ChangeNotifier {
 
   void loadBuildings() {
     // todo: parse data from simple json file not dart class
-    CampusPolygons.buildings.forEach((key, value) {
+    CampusBuildings.buildings.forEach((key, value) {
       buildings.add(new Building(key, value));
     });
   }
