+++ lib/screens/home_page.dart
@@ -1,8 +1,9 @@
 import 'package:concordia_navigation/widgets/custom_drawer.dart';
 import 'package:flutter/material.dart';
-import 'package:concordia_navigation/widgets/custom_appbar.dart';
+import 'package:concordia_navigation/widgets/homepage_appbar.dart';
 import 'package:concordia_navigation/widgets/map_widget.dart';
 
+//The app will launch here.
 class HomePage extends StatefulWidget {
   @override
   _HomePageState createState() => _HomePageState();
@@ -12,7 +13,7 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: CustomAppBar(),
+      appBar: HomePageAppBar(),
       drawer: CustomDrawer(),
       body: MapWidget(),
     );
