+++ lib/screens/profile.dart
@@ -1,6 +1,8 @@
+import 'package:concordia_navigation/storage/app_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:concordia_navigation/services/localization.dart';
 
+//User Profile Screen
 class Profile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -9,15 +11,15 @@ class Profile extends StatelessWidget {
         title: Text(
           ConcordiaLocalizations.of(context).profile,
         ),
-        backgroundColor: Color(0xFF73C700),
+        backgroundColor: greenColor,
       ),
       body: Center(
         child: RaisedButton(
           child: Text(
             ConcordiaLocalizations.of(context).profile,
           ),
-          color: Color(0xFF73C700),
-          textColor: Colors.white,
+          color: greenColor,
+          textColor: whiteColor,
           onPressed: () {
             showDialog(
                 context: context,
