+++ lib/screens/settings.dart
@@ -1,6 +1,8 @@
+import 'package:concordia_navigation/storage/app_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:concordia_navigation/services/localization.dart';
 
+//App Settings Screen
 class Settings extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -9,13 +11,13 @@ class Settings extends StatelessWidget {
         title: Text(
           ConcordiaLocalizations.of(context).settings,
         ),
-        backgroundColor: Color(0xFF73C700),
+        backgroundColor: greenColor,
       ),
       body: Center(
         child: RaisedButton(
           child: Text(ConcordiaLocalizations.of(context).settings),
-          color: Color(0xFF73C700),
-          textColor: Colors.white,
+          color: greenColor,
+          textColor: whiteColor,
           onPressed: () {
             showDialog(
                 context: context,
