+++ lib/screens/shuttle_schedule.dart
@@ -1,6 +1,8 @@
+import 'package:concordia_navigation/storage/app_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:concordia_navigation/services/localization.dart';
 
+//Concordia Shuttle Schedule Screen
 class ShuttleSchedule extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -9,7 +11,7 @@ class ShuttleSchedule extends StatelessWidget {
         title: Text(
           ConcordiaLocalizations.of(context).shuttle,
         ),
-        backgroundColor: Color(0xFF73C700),
+        backgroundColor: greenColor,
       ),
       body: DataTable(
         columns: [
