+++ lib/services/directions_service.dart
@@ -1,12 +1,13 @@
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-
 import 'network.dart';
 
+//Google API key
+//TODO: Hide Key
 const String apiKey = 'AIzaSyBHXKzGZEeBhP_m3QQ6vpI0hRODxeeEWl0';
 
+//This class will fetch the JSON from URL
 class DirectionsService {
-  static final String mapsUrl =
-      'maps.googleapis.com/maps/api/directions/json'; //origin=45.4957,%20-73.5781&destination=45.4582,%20-73.6405';
+  static final String mapsUrl = 'maps.googleapis.com/maps/api/directions/json';
 
   static Future<String> getDirections(final LatLng start, final LatLng end,
       {String mode = "driving"}) async {
