+++ lib/services/location_search.dart
@@ -1,17 +1,19 @@
+import 'package:concordia_navigation/storage/app_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:concordia_navigation/providers/map_data.dart';
 
+/*This class extends Search Delegate class implemented by flutter.
+It will be called when the user clicks on the search button in the Appbar.
+*/
 class LocationSearch extends SearchDelegate {
-  LatLng sgw = LatLng(45.495944, -73.578075);
-  LatLng loyola = LatLng(45.4582, -73.6405);
-
+  ///This method returns suggested locations to the user, in this case Loyola and SGW campus.
   @override
   Widget buildSuggestions(BuildContext context) {
     return ListView(
       children: <Widget>[
         Consumer<MapData>(
+          ///Wrapped in Consumer, listening to Provider **ConcreteObserver**
           builder: (context, mapData, child) {
             return ListTile(
               leading: Icon(Icons.location_city),
@@ -20,18 +22,18 @@ class LocationSearch extends SearchDelegate {
               onTap: () {
                 Navigator.of(context).pop();
                 mapData.controllerDestination.text = "SGW, Montreal";
-                mapData.controllerStaring.text = "Current Location";
+                mapData.controllerStarting.text = "Current Location";
                 mapData.changeStart(mapData.getCurrentLocation);
                 mapData.changeCampus('sgw');
                 mapData.changeEnd(sgw);
                 mapData.changeMode("driving");
-
                 Navigator.pushNamed(context, '/directions');
               },
             );
           },
         ),
         Consumer<MapData>(
+          ///Wrapped in Consumer, listening to Provider **ConcreteObserver**
           builder: (context, mapData, child) {
             return ListTile(
               leading: Icon(Icons.location_city),
@@ -40,12 +42,11 @@ class LocationSearch extends SearchDelegate {
               onTap: () {
                 Navigator.of(context).pop();
                 mapData.controllerDestination.text = "Loyola Campus, Montreal";
-                mapData.controllerStaring.text = "Current Location";
+                mapData.controllerStarting.text = "Current Location";
                 mapData.changeStart(mapData.getCurrentLocation);
                 mapData.changeCampus('loyola');
                 mapData.changeEnd(loyola);
                 mapData.changeMode("driving");
-
                 Navigator.pushNamed(context, '/directions');
               },
             );
@@ -55,6 +56,7 @@ class LocationSearch extends SearchDelegate {
     );
   }
 
+  ///This method adds a return IconButton to return to the homepage.
   @override
   Widget buildLeading(BuildContext context) {
     return IconButton(
@@ -67,6 +69,7 @@ class LocationSearch extends SearchDelegate {
         });
   }
 
+  ///This method add the a clear IconButton to clear user's input.
   @override
   List<Widget> buildActions(BuildContext context) {
     return [
@@ -79,6 +82,8 @@ class LocationSearch extends SearchDelegate {
     ];
   }
 
+  ///This method returns results from user input.
+  ///To be implemented with Google Places.
   @override
   Widget buildResults(BuildContext context) {
     return Container();
