+++ lib/services/shuttle_data.dart
@@ -3,9 +3,11 @@ import 'package:flutter/foundation.dart';
 import 'dart:async' show Future;
 import 'package:flutter/services.dart' show rootBundle;
 
+//This class is used to parse the shuttle schedule JSON and get the next available shuttle bus.
 class ShuttleData extends ChangeNotifier {
   var _json;
 
+  ///Read JSON when shuttle data is requested
   ShuttleData() {
     readJsonFile();
   }
