+++ lib/storage/app_constants.dart
@@ -1,0 +1,19 @@
+///All application constants are stored here.
+
+import 'package:flutter/cupertino.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+
+const LatLng sgw = LatLng(45.495944, -73.578075);
+const LatLng loyola = LatLng(45.4582, -73.6405);
+
+const double CAMERA_ZOOM = 16;
+const double CAMERA_TILT = 50;
+const double CAMERA_BEARING = 30;
+
+const Color greenColor = Color(0xFF73C700);
+const Color whiteColor = Color(0xFFFFFFFF);
+const Color blueColor = Color(0xFF017BFF);
+const Color blackColor = Color(0xFF000000);
+const Color offWhiteColor = Color(0xFFFFFFF8);
+const Color lightGreyColor = Color(0xFFF0F0F0);
+const Color greyColor = Color(0xFF656363);
