+++ lib/widgets/homepage_appbar.dart
@@ -1,17 +1,20 @@
 import 'package:concordia_navigation/providers/buildings_data.dart';
+import 'package:concordia_navigation/storage/app_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:concordia_navigation/services/location_search.dart';
 import 'package:provider/provider.dart';
 
-class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
+//Custom AppBar shown only in HomePage
+class HomePageAppBar extends StatelessWidget implements PreferredSizeWidget {
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      backgroundColor: Color(0xFF73C700),
+      backgroundColor: greenColor,
       title: Text("ConNavigation"),
       actions: <Widget>[
         Consumer<BuildingsData>(
-          builder: (BuildContext context, buildings, Widget child) => IconButton(
+          builder: (BuildContext context, buildings, Widget child) =>
+              IconButton(
             icon: Icon(Icons.wb_sunny),
             onPressed: () {
               if (buildings.visible) {
